spark8s/requirements.txt
delta-spark==3.1.0
google-cloud-storage